requirements.txt
discord
python-dotenv
deep_translator
